requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.3.0
diff-match-patch==20121119
Django==1.10.8
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==1.1.0
django-formtools==2.1
django-import-export==1.1.0
django-reversion==2.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.9.2
idna==2.7
ipython==7.1.1
ipython-genutils==0.2.0
itypes==1.1.0
jdcal==1.4
jedi==0.13.1
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
odfpy==1.3.6
openpyxl==2.5.8
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.7
ptyprocess==0.6.0
Pygments==2.2.0
PyJWT==1.6.4
pytz==2018.7
PyYAML==3.13
requests==2.20.1
six==1.11.0
tablib==0.12.1
traitlets==4.3.2
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
xlrd==1.1.0
XlsxWriter==1.1.2
xlwt==1.3.0
